Expose Die constructors publicly
Originally I wanted consumers of this class to only parse a die
description using the `die` function, but that is cumbersome to work
with in practice, so I decided to expose the Die constructors so that
Dies can be created directly.

To facilitate testing, I also derived Eq from Die, which will probably
be beneficial outside of testing as well.