Define Rollable behavior
A Rollable is anything that can produce an integral result. Currently
that is dice (represented by the Die type), but eventually we will
incorporate constant values when we handle roll descriptions like
"2d4+10". The constant values can essentially be treated as dice rolls
that always return the same value; this concept is encapsulated by the
Rollable typeclass.